+++ src/Proto3/Suite/DotProto/Rendering.hs
@@ -1,11 +1,9 @@
 -- | This module provides types and functions to generate .proto files.
 
-{-# LANGUAGE CPP #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE OverloadedLists            #-}
 {-# LANGUAGE RecordWildCards            #-}
 {-# LANGUAGE ScopedTypeVariables        #-}
-{-# OPTIONS_GHC -fno-warn-orphans       #-}
 
 module Proto3.Suite.DotProto.Rendering
   ( renderDotProto
@@ -19,17 +17,11 @@ module Proto3.Suite.DotProto.Rendering
   , Pretty(..)
   ) where
 
-import           Data.Char
-import qualified Data.List.NonEmpty              as NE
-#if (MIN_VERSION_base(4,11,0))
-import           Prelude                         hiding ((<>))
-#endif
 import           Proto3.Suite.DotProto.AST
 import           Proto3.Wire.Types               (FieldNumber (..))
-import           Text.PrettyPrint                (($$), (<+>), (<>))
+import           Text.PrettyPrint                (($$), (<+>))
 import qualified Text.PrettyPrint                as PP
 import           Text.PrettyPrint.HughesPJClass  (Pretty(..))
-import           Turtle.Compat                   (encodeString)
 
 -- | Options for rendering a @.proto@ file.
 data RenderingOptions = RenderingOptions
@@ -67,30 +59,6 @@ renderDotProto opts DotProto{..}
  $$ (PP.vcat $ topOption <$> protoOptions)
  $$ (PP.vcat $ prettyPrintProtoDefinition opts <$> protoDefinitions)
 
-strLit :: String -> PP.Doc
-strLit string = PP.text "\"" <> foldMap escape string <> PP.text "\""
-  where
-    escape '\n' = PP.text "\\n"
-    escape '\\' = PP.text "\\\\"
-    escape '\0' = PP.text "\\x00"
-    escape '"'  = PP.text "\\\""
-    escape  c   = PP.text [ c ]
-
-instance Pretty DotProtoPackageSpec where
-  pPrint (DotProtoPackageSpec p) = PP.text "package" <+> pPrint p <> PP.text ";"
-  pPrint (DotProtoNoPackage)     = PP.empty
-
-instance Pretty DotProtoImport where
-  pPrint (DotProtoImport q i) =
-    PP.text "import" <+> pPrint q <+> strLit fp <> PP.text ";"
-    where
-      fp = encodeString i
-
-instance Pretty DotProtoImportQualifier where
-  pPrint DotProtoImportDefault = PP.empty
-  pPrint DotProtoImportPublic  = PP.text "public"
-  pPrint DotProtoImportWeak    = PP.text "weak"
-
 optionAnnotation :: [DotProtoOption] -> PP.Doc
 optionAnnotation [] = PP.empty
 optionAnnotation os = PP.brackets
@@ -99,10 +67,7 @@ optionAnnotation os = PP.brackets
                     $ pPrint <$> os
 
 topOption :: DotProtoOption -> PP.Doc
-topOption o = PP.text "option" <+> pPrint o <> PP.text ";"
-
-instance Pretty DotProtoOption where
-  pPrint (DotProtoOption key value) = pPrint key <+> PP.text "=" <+> pPrint value
+topOption o = PP.text "option" <+> pPrint o PP.<> PP.text ";"
 
 renderComment :: String -> PP.Doc
 renderComment = PP.vcat . map ((PP.text "//" <+>) . textIfNonempty) . lines
@@ -133,19 +98,19 @@ prettyPrintProtoDefinition opts = defn where
   msgPart _       (DotProtoMessageReserved reservations)
     =   PP.text "reserved"
     <+> (PP.hcat . PP.punctuate (PP.text ", ") $ pPrint <$> reservations)
-    <>  PP.text ";"
+    PP.<> PP.text ";"
   msgPart msgName (DotProtoMessageOneOf name fields)     = vbraces (PP.text "oneof" <+> pPrint name) (PP.vcat $ field msgName <$> fields)
   msgPart _       (DotProtoMessageOption opt)
-    = PP.text "option" <+> pPrint opt <> PP.text ";"
+    = PP.text "option" <+> pPrint opt PP.<> PP.text ";"
 
   field :: DotProtoIdentifier -> DotProtoField -> PP.Doc
   field msgName (DotProtoField number mtype name options comments)
     =   pPrint mtype
     <+> roSelectorName opts msgName name number
     <+> PP.text "="
-    <+> pPrint number
+    <+> pPrintFieldNumber number
     <+> optionAnnotation options
-    <>  PP.text ";"
+    PP.<> PP.text ";"
     $$  PP.nest 2 (renderComment comments)
 
   enumPart :: DotProtoIdentifier -> DotProtoEnumPart -> PP.Doc
@@ -154,72 +119,14 @@ prettyPrintProtoDefinition opts = defn where
     <+> PP.text "="
     <+> pPrint (fromIntegral value :: Int)
     <+> optionAnnotation options
-    <> PP.text ";"
-  enumPart _       (DotProtoEnumReserved reservedFields) 
+    PP.<> PP.text ";"
+  enumPart _       (DotProtoEnumReserved reservedFields)
     = PP.text "reserved" <+> (PP.hcat . PP.punctuate (PP.text ", ") $ pPrint <$> reservedFields)
   enumPart _       (DotProtoEnumOption opt)
-    = PP.text "option" <+> pPrint opt <> PP.text ";"
+    = PP.text "option" <+> pPrint opt PP.<> PP.text ";"
 
-instance Pretty DotProtoServicePart where
-  pPrint (DotProtoServiceRPCMethod RPCMethod{..})
-    =   PP.text "rpc"
-    <+> pPrint rpcMethodName
-    <+> PP.parens (pPrint rpcMethodRequestStreaming <+> pPrint rpcMethodRequestType)
-    <+> PP.text "returns"
-    <+> PP.parens (pPrint rpcMethodResponseStreaming <+> pPrint rpcMethodResponseType)
-    <+> case rpcMethodOptions of
-          [] -> PP.text ";"
-          _  -> PP.braces . PP.vcat $ topOption <$> rpcMethodOptions
-  pPrint (DotProtoServiceOption option) = topOption option
-
-instance Pretty Streaming where
-  pPrint Streaming    = PP.text "stream"
-  pPrint NonStreaming = PP.empty
-
-instance Pretty DotProtoIdentifier where
-  pPrint (Single name)                    = PP.text name
-  pPrint (Dots (Path names))              = PP.hcat . PP.punctuate (PP.text ".") $ PP.text <$> NE.toList names
-  pPrint (Qualified qualifier identifier) = PP.parens (pPrint qualifier) <> PP.text "." <> pPrint identifier
-  pPrint Anonymous                        = PP.empty
-
-instance Pretty DotProtoValue where
-  pPrint (Identifier value) = pPrint value
-  pPrint (StringLit  value) = strLit value
-  pPrint (IntLit     value) = PP.text $ show value
-  pPrint (FloatLit   value) = PP.text $ show value
-  pPrint (BoolLit    value) = PP.text $ toLower <$> show value
-
-instance Pretty DotProtoType where
-  pPrint (Prim           ty) = pPrint ty
-  pPrint (Repeated       ty) = PP.text "repeated" <+> pPrint ty
-  pPrint (NestedRepeated ty) = PP.text "repeated" <+> pPrint ty
-  pPrint (Map keyty valuety) = PP.text "map<" <> pPrint keyty <> PP.text ", " <> pPrint valuety <> PP.text ">"
-
-instance Pretty DotProtoPrimType where
-  pPrint (Named i)  = pPrint i
-  pPrint Int32      = PP.text "int32"
-  pPrint Int64      = PP.text "int64"
-  pPrint SInt32     = PP.text "sint32"
-  pPrint SInt64     = PP.text "sint64"
-  pPrint UInt32     = PP.text "uint32"
-  pPrint UInt64     = PP.text "uint64"
-  pPrint Fixed32    = PP.text "fixed32"
-  pPrint Fixed64    = PP.text "fixed64"
-  pPrint SFixed32   = PP.text "sfixed32"
-  pPrint SFixed64   = PP.text "sfixed64"
-  pPrint String     = PP.text "string"
-  pPrint Bytes      = PP.text "bytes"
-  pPrint Bool       = PP.text "bool"
-  pPrint Float      = PP.text "float"
-  pPrint Double     = PP.text "double"
-
-instance Pretty FieldNumber where
-  pPrint = PP.text . show . getFieldNumber
-
-instance Pretty DotProtoReservedField where
-  pPrint (SingleField num)      = PP.text $ show num
-  pPrint (FieldRange start end) = (PP.text $ show start) <+> PP.text "to" <+> (PP.text $ show end)
-  pPrint (ReservedIdentifier i) = PP.text $ show i
+pPrintFieldNumber :: FieldNumber -> PP.Doc
+pPrintFieldNumber = PP.text . show . getFieldNumber
 
 -- | Render protobufs metadata as a .proto file stringy
 toProtoFile :: RenderingOptions -> DotProto -> String